Refactor spacetime: Merge code to fewer files
Merge all code in spacetime to only core.[hc]pp and io.[hc]pp.
