+++ cpp/binary/pymod_modmesh/CMakeLists.txt
@@ -14,6 +14,7 @@ pybind11_add_module(
     ${MODMESH_HEADERS}
     ${MODMESH_SOURCES}
     ${MODMESH_SPACETIME_HEADERS}
+    ${MODMESH_SPACETIME_SOURCES}
     ${MODMESH_PYTHON_HEADERS}
     ${MODMESH_PYTHON_SOURCES}
 )
