+++ cpp/binary/viewer/CMakeLists.txt
@@ -33,6 +33,7 @@ qt_add_executable(
     ${MODMESH_VIEW_HEADERS}
     ${MODMESH_VIEW_SOURCES}
     ${MODMESH_SPACETIME_HEADERS}
+    ${MODMESH_SPACETIME_SOURCES}
     ${MODMESH_PYTHON_HEADERS}
     ${MODMESH_PYTHON_SOURCES}
     ${MODMESH_PYMOD_MODMESH_HEADERS}
