+++ cpp/modmesh/spacetime/CMakeLists.txt
@@ -4,24 +4,17 @@
 cmake_minimum_required(VERSION 3.16)
 
 set(MODMESH_SPACETIME_HEADERS
-    ${CMAKE_CURRENT_SOURCE_DIR}/Celm.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Celm_decl.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/ElementBase.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/ElementBase_decl.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Field.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Field_decl.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Grid.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Grid_decl.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/core.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/io.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Selm.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Selm_decl.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Solver.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/SolverBase.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/SolverBase_decl.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/spacetime.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/base_spacetime.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/kernel/inviscid_burgers.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/kernel/linear_scalar.hpp
+    CACHE FILEPATH "" FORCE
+    )
+
+set(MODMESH_SPACETIME_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/io.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/core.cpp
     CACHE FILEPATH "" FORCE
 )
