+++ cpp/modmesh/spacetime/io.hpp
@@ -7,87 +7,24 @@
 
 #include <iostream>
 
-#include <modmesh/spacetime/Grid.hpp>
-#include <modmesh/spacetime/Celm.hpp>
-#include <modmesh/spacetime/Solver.hpp>
-#include <modmesh/spacetime/Selm.hpp>
+#include <modmesh/spacetime/core.hpp>
 #include <modmesh/spacetime/kernel/linear_scalar.hpp>
 #include <modmesh/spacetime/kernel/inviscid_burgers.hpp>
 
 namespace spacetime
 {
 
-inline std::ostream & operator<<(std::ostream & os, const Grid & grid)
-{
-    os << "Grid(xmin=" << grid.xmin() << ", xmax=" << grid.xmax() << ", ncelm=" << grid.ncelm() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const Field & sol)
-{
-    os << "Field(grid=" << sol.grid() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const Solver & sol)
-{
-    os << "Solver(grid=" << sol.grid() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const Celm & elm)
-{
-    os << "Celm(" << (elm.on_even_plane() ? "even" : "odd") << ", ";
-    os << "index=" << elm.index() << ", x=" << elm.x() << ", xneg=" << elm.xneg() << ", xpos=" << elm.xpos() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const Selm & elm)
-{
-    os << "Selm(" << (elm.on_even_plane() ? "even" : "odd") << ", ";
-    os << "index=" << elm.index() << ", x=" << elm.x() << ", xneg=" << elm.xneg() << ", xpos=" << elm.xpos() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const InviscidBurgersSolver & sol)
-{
-    os << "InviscidBurgersSolver(grid=" << sol.grid() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const InviscidBurgersCelm & elm)
-{
-    os << "InviscidBurgersCelm(" << (elm.on_even_plane() ? "even" : "odd") << ", ";
-    os << "index=" << elm.index() << ", x=" << elm.x() << ", xneg=" << elm.xneg() << ", xpos=" << elm.xpos() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const InviscidBurgersSelm & elm)
-{
-    os << "InviscidBurgersSelm(" << (elm.on_even_plane() ? "even" : "odd") << ", ";
-    os << "index=" << elm.index() << ", x=" << elm.x() << ", xneg=" << elm.xneg() << ", xpos=" << elm.xpos() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const LinearScalarSolver & sol)
-{
-    os << "LinearScalarSolver(grid=" << sol.grid() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const LinearScalarCelm & elm)
-{
-    os << "LinearScalarCelm(" << (elm.on_even_plane() ? "even" : "odd") << ", ";
-    os << "index=" << elm.index() << ", x=" << elm.x() << ", xneg=" << elm.xneg() << ", xpos=" << elm.xpos() << ")";
-    return os;
-}
-
-inline std::ostream & operator<<(std::ostream & os, const LinearScalarSelm & elm)
-{
-    os << "LinearScalarSelm(" << (elm.on_even_plane() ? "even" : "odd") << ", ";
-    os << "index=" << elm.index() << ", x=" << elm.x() << ", xneg=" << elm.xneg() << ", xpos=" << elm.xpos() << ")";
-    return os;
-}
+std::ostream & operator<<(std::ostream & os, const Grid & grid);
+std::ostream & operator<<(std::ostream & os, const Field & sol);
+std::ostream & operator<<(std::ostream & os, const Solver & sol);
+std::ostream & operator<<(std::ostream & os, const Celm & elm);
+std::ostream & operator<<(std::ostream & os, const Selm & elm);
+std::ostream & operator<<(std::ostream & os, const InviscidBurgersSolver & sol);
+std::ostream & operator<<(std::ostream & os, const InviscidBurgersCelm & elm);
+std::ostream & operator<<(std::ostream & os, const InviscidBurgersSelm & elm);
+std::ostream & operator<<(std::ostream & os, const LinearScalarSolver & sol);
+std::ostream & operator<<(std::ostream & os, const LinearScalarCelm & elm);
+std::ostream & operator<<(std::ostream & os, const LinearScalarSelm & elm);
 
 } /* end namespace spacetime */
